Refactor parse_tweet_objects.py to parse_tweet_objects package.
Rearrange source code according to functionality into three
modules: parser, tweet, utils.

Implement type hinting.

Implement basic logging. Replace calls to print() and
warnings.warn() with logging calls.

Replace old-style string formatting with f-strings.

Clean up doc-strings. Add some comments clarifying logic.

Fix ParseTweetsFromJSONL __init__ method.

Remove __main__ function.
